+++ fstone/config/flight-stone.json
@@ -1,3 +1,4 @@
 {
-    "default_system_device": "Kinect",
+    "default_patch_name": "ddddddd",
+    "default_system_device": "Kinect"
 }
